python-black/week12/Class12K_sukhanov_artem/requirements.txt
aiohttp~=3.11.10
aiofiles~=24.1.0
beautifulsoup4~=4.12.3
oauth2client~=4.1.3
python-dotenv~=1.0.1
google-api-python-client~=2.154.0